register dummy macos_exporter integration on non-Darwin systems
