+++ pkg/integrations/node_exporter/config.go
@@ -321,19 +321,3 @@ func (f *flags) addBools(m map[*bool]string) {
 		}
 	}
 }
-
-func init() {
-	// Register macos_exporter
-	integrations.RegisterIntegration(&DarwinConfig{})
-}
-
-// DarwinConfig extends the Config struct and overrides the name of
-// the integration to avoid conflicts with node_exporter integration.
-type DarwinConfig struct {
-	Config
-}
-
-// Name returns the name of the integration that this config represents.
-func (*DarwinConfig) Name() string {
-	return "macos_exporter"
-}
